updating version of material + new cards reusable components
