+++ src/components/PoolCardContent/index.ts
@@ -1,0 +1,1 @@
+export {default} from './PoolCardContent';
